[MOD] added alert/:id page, added error checking to account creation
